Do not remove plugin resources on operator uninstall. (#162)
When removing the CSV, OLM can remove controller's role or rolebinding
before raising the SIGTERM on the controller, which makes the controller
to fail to delete the plugin resources due to lack of permissions.

This is a workaround for some issues related to plugin's resources when
uninstalling the operator. A better approach for plugin
install/uninstall is yet to implement.

To fully disable/uninstall the operator plugin, it must be removed
manually after the operator has been uninstalled.

If they're not removed and the operator is installed again, the
resources are applied but the error is ignored to let the controller
continue working normally.

Also:
- Minor changes to controller RBAC, to have just one resource per
verb list. Otherwise, the same resource can appear in several verbs list
making it a bit hard to follow.
- Added logging traces to the console plugin resources
creation/removal.